--- a/src/afterfact.rs
+++ b/src/afterfact.rs
@@ -1145,6 +1145,13 @@ fn output_json_str(
         "TgtCountry",
         "TgtCity",
     ];
+    let key_add_to_details_cnt = key_add_to_details.iter().filter(|target_key| {
+        ext_field_map
+            .get(&CompactString::from(**target_key))
+            .unwrap()
+            .to_value()
+            != "-"
+    }).count();
     for (key, profile) in ext_field.iter() {
         let val = profile.to_value();
         let vec_data = _get_json_vec(profile, &val.to_string());
@@ -1246,13 +1253,7 @@ fn output_json_str(
                             let key = _convert_valid_json_str(&[output[0]], false);
                             let fmted_val = _convert_valid_json_str(&output, false);
                             let last_contents_end = if is_included_geo_ip
-                                && key_add_to_details.iter().any(|target_key| {
-                                    ext_field_map
-                                        .get(&CompactString::from(*target_key))
-                                        .unwrap()
-                                        .to_value()
-                                        != "-"
-                                }) {
+                                && key_add_to_details_cnt > 0 {
                                 ","
                             } else {
                                 ""
@@ -1271,7 +1272,8 @@ fn output_json_str(
                         }
                     }
                     if is_included_geo_ip {
-                        for (suffix_cnt, target_key) in key_add_to_details.iter().enumerate() {
+                        let mut add_cnt = 0;
+                        for target_key in key_add_to_details.iter() {
                             let val = ext_field_map
                                 .get(&CompactString::from(*target_key))
                                 .unwrap()
@@ -1279,7 +1281,8 @@ fn output_json_str(
                             if val == "-" {
                                 continue;
                             }
-                            let output_end_fmt = if suffix_cnt == key_add_to_details.len() - 1 {
+                            add_cnt += 1;
+                            let output_end_fmt = if key_add_to_details_cnt == add_cnt {
                                 ""
                             } else {
                                 ","