@hitenkoku 2023-01-29
in https://github.com/Yamato-Security/hayabusa/pull/896/commits/6505d90b9d3c5788a162cbd342c6f6059d9f26ab
 
|Version| Elapsed time | Memory(peak)  | Events with hits / Total events | Output file size(bytes)|
|-------| ------------ | --------|--------|--------| 
| [before](https://github.com/Yamato-Security/hayabusa/commit/b97b22403eb4ef4633b7e5a32b499051ae7e395e)  | 00:03:52.226| 4.8 GiB | 1,593,715 / 4,817,181 | 575078117 |
| [This PR](https://github.com/Yamato-Security/hayabusa/pull/896/commits/2d89759ab5e56c16a69e6ba7b8fc6641377b5783)   |  00:03:59.750| 5.7 GiB | 1,593,715 / 4,817,181 | 614236764 | 
| https://github.com/Yamato-Security/hayabusa/pull/896/commits/6505d90b9d3c5788a162cbd342c6f6059d9f26ab | 00:03:57.280  | 5.7GiB | 1,593,715 / 4,817,181 | 614236764 |