@hitenkoku 2023-01-29
@YamatoSecurity  @fukusuket  moved ASN, City, and Country in details field when json-timeline is specified in https://github.com/Yamato-Security/hayabusa/pull/896/commits/693ee7dc3d5f6e58e4a50ce50ff34fe1d404dcb9 .