@YamatoSecurity 2023-01-30
Humm.. I tried both CSV and JSON timelines but don't see any of the additional MaxMind fields. Can you check that it is working after the latest commits?