@hitenkoku 2023-01-30
I checked existed SrcASN etc...  in json and csv in latest commit.
Would you share executed command and environment?

> Humm.. I tried both CSV and JSON timelines but don't see any of the additional MaxMind fields. Can you check that it is working after the latest commits?