@hitenkoku 2023-01-30
@YamatoSecurity I fixed following problem in https://github.com/Yamato-Security/hayabusa/pull/896/commits/1a02aaf508ecfe85c2c4d607aeb08a394b739e60

Would you check it?

> ```
> ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> [ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> [ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> [ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> [ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> thread 'hayabusa-thread[ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> ' panicked at 'thread 'called `Option::unwrap()` on a `None` valuehayabusa-thread', ' panicked at 'src\detections\message.rscalled `Option::unwrap()` on a `None` value:', 146src\detections\message.rs::21146
> :note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
> 21
> thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value: JoinError::Panic(Id(1077978), ...)', src\detections\detection.rs:157:41
> ```
> 
> Can you check this?



> @hitenkoku Thank you so much the PR! On my dataset, I get the following error when scanning:
> 
> ```
> ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> [ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> [ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> [ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> [ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> thread 'hayabusa-thread[ERROR] IoError("Failed Convert IP Address. input: LOCAL")
> ' panicked at 'thread 'called `Option::unwrap()` on a `None` valuehayabusa-thread', ' panicked at 'src\detections\message.rscalled `Option::unwrap()` on a `None` value:', 146src\detections\message.rs::21146
> :note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
> 21
> thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value: JoinError::Panic(Id(1077978), ...)', src\detections\detection.rs:157:41
> ```
> 
> Can you check this?